requirements.txt
asgiref==3.11.0
Django==5.2.10
pillow==12.1.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3


Django==5.2.10
Pillow==10.2.0
gunicorn==21.2.0
